Move tests to their relevant package.
