+++ main/tests/server/src/com/google/refine/operations/column/UrlFetchingTests.java
@@ -31,7 +31,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 */
 
-package com.google.refine.tests.model;
+package com.google.refine.operations.column;
 
 import java.io.IOException;
 import java.net.InetAddress;
